BUG: Filter row with ages that are marked as "Don't Know" or "Refused" to prevent them from being classified as stillbirths.
This fix creates an additional intermediate file "invalid-age-prepped.csv" with
the filtered rows. It also prioritizes data which is present in cases which the
module is marked as missing but there is data. It doesn't catch invalid values
of the module variable, only string missing and valid missing flags (8 and 9).