MAINT: Separate scoring of validation and user data
Scoring based on symptoms is moved to a separate method. This is the
common machinery for scoring each row. User data records for the user
data also store restrictions and rule based predictions. The validation
data stores the gold standard cause and should be duplicated based on
the uniform resampling frequencies.