ENH: Add output file with multiple predictions
All non-masked predictions are given a likelihood based on the range of
ranks for the corresponding cause. Likelihoods are determined after
predicting and rule-based predictions are given the highest likelihood.
The app now outputs a file '{module}-likelihoods.csv' in the root of the
output dir with predictions and csmfs.