Add censoring rule for stillbirth if age>0
Users find it implausible to call a death a stillbirth if the age
at death is greater than zero. We already have a rule for
determining if something is stillbirth. This adds exclusions to
help determine if something isn't. However, the logic rules currently
take precedence over the censorings, which are applied at tariff
level. This means if that rule is endorsed the age censoring is
irrelevant.