ENV: Remove matplotlibrc
Currently the modules that use matplotlib explicitly call `use`
passing the 'Agg' backend. The rc file in the project root
provides a fallback if `use` is not called. However, the file is
not packaged with the executable so the presence of the fallback
setting creates a dev and test environment that is different from
the runtime environment and could lead to graphing bugs. It is
preferable to explicitly set backend with `use` because then it
cannot be overrided with an environment variable, whereas the
packaged matplotlibrc file could be overrided. Also, we do want
'Agg', which is a non-interactive backend for writing PNGs,
instead of 'TKAgg' or 'WXAgg', which are interactive backends for
using matplotlib in GUI applications.