BUG: Use safe_float to check injury duration
If the injury duration column is missing or has the value of '' we
want to assume the injury was recent enough to be considered as the
cause of death. Using safe_float returns a zero which is less than
the cutoff. This now treats invalid entries as missing and thus can
lead to an endorsement.