BUG: Treat out of range multiselect values as invalid
Previously we checked that a multiselect value could be split and
cast to ints. If errors occurs we treated the whole string as
invalid. Now we are also considering it invalid if it contains any
integer value that doesn't appear in the codebook, even if some of
the values are valid.