BUG: Ensure duration thresholds under 1 are truncated
We use greater than or equal for endorsing duration symptoms. The
input data from ODK must be integers which are scaled to days and
written to an intermediate file. Serializing leads to rounding
error for certain thresholds. We want to ensure that the threshold
value (e.g. 5 hours) leads to an endorsement by testing against a
truncated threshold.