ENH: Exit CLI app with the exit status.
The CLI is wrapped in a click.Command which exits successfully
even if the thread dies with an exception. We now cache the
completion status from the thread and exit the command using the
same exit code. This allows other command line tools that are
looking for a non-zero exit status to function properly.