ENH: Change residual ICD causes
Reassign 'Other' category ICDs that cause confusion with other
systems using this data.

From Sonja-We have used R100 and R101 to represent 'Other NCD' and
'Other defined causes of childhood illness' respectively. However,
when we put this information into ANACONDA (for instance)
it converts it to R10.0 Acute abdomen and R10.1 Pain localised in
the upper abdomen.